requirements.txt
loguru==0.5.3
tqdm==4.50.0
retrying==1.3.3
requests==2.23.0
scikit-learn==0.23.1
numpy==1.19.2
pyyaml==5.4.1
regex==2.5.93